Simplify the output of the account usage information
I'm not that familiar with Laravel and all the ins and outs of its
various parts and components. Given that, I didn't know just how simple
outputting information via Laravel Blade templates and components could
be. Now that I've learned more about it, I'm making this commit to
remove the extra, unnecessary work required to achieve the same result
as in the previous commit.